global: refactor. remove utils file and create classes
